// src/generated/client/worldmonitor/market/v1/service_client.ts
export interface ListGulfQuotesResponse {
  quotes: GulfQuote[];
  rateLimited: boolean;
}

export interface GulfQuote {
  symbol: string;
  name: string;
  flag: string;
  country: string;
  type: string;
  price: number;
  change: number;
  sparkline: number[];
}